initial content
The code is a copy of the csi-common and hostpath code from
github.com/kubernetes-csi/drivers. csi-common is expected to become
deprecated, therefore we copy instead of import the code.

The "hostpath" and "CSI common" names were replaced, without
changing any of the original functionality.

"make push-oio-csi-driver" and "make test" pass.