SPDK: fix error response handling
When receiving an error map, ReadResponseHeader returned an error
instead of using the error information to prepare an error string for
the response. As a result, the connection was considered in a bad
state. Now the map is turned into a normal string.

Because some client code will have to check for well-known error
codes, IsJSONError decodes such strings and compares the error code.