OIO controller: map malloc volume
MapVolume for a malloc volume creates and activates it, i.e. it
combines provisioning and publishing (in CSI terms). For unit testing
that is good enough. In production, provisioning will be triggered by
the OIO CSI driver or some other central component, and then MapVolume
only needs to activate the existing volume.