Makefile: split up .PHONY and testing
This removes the central line with .PHONY in favor of setting that
where the target is defined. This way changes to the Makefile are more
localized.

"tests" is now a meta target that triggers all tests. Actually running
tests is done by "run_tests". This has the advantage that one can skip
the build testing if desired.