OIO controller: identify block devices
The virtual machine must be able to determine which of its current
block devices corresponds to the request volume. The SPDK developers
suggested to use the PCI address and SCSI target +
LUN (https://github.com/spdk/spdk/issues/330).

A slightly different approach is used to avoid complex PCI topology
parsing: the administrator has to know how the Linux kernel encodes
the PCI topology in /sys/dev/block and give the OIO controller the
resulting substring. That then gets returned to the OIO CSI driver
which can find the right block device with a simple substring search.