SPDK: build binary from vendored source + use for testing
Vendoring in the SPDK/DPDK source code relies on a version of the dep
tool that exports
submodules (https://github.com/golang/dep/pull/1909).  In addition,
there must be at least one Go source file in the SPDK
repo (https://github.com/golang/dep/issues/1306), so we have to use a
copy of the SPDK repo which
does (https://review.gerrithub.io/#/c/spdk/spdk/+/416831).

Overall, including the source this way is experimental. The
alternative is to use git submodules.