Makefile: clean up test configuration
Having to use TEST_QEMU_IMAGE=clear-kvm without the .img suffix and
without the directory was awkward and inconsistent with
TEST_SPDK_VHOST_BINARY. We can derive the helper files also when the
actual file is given, so that is how it now needs to be specified.

WITH_E2E_TESTS=1 or any other non-empty value will set the defaults,
but only when nothing else has been manually selected.