QEMU: port forwarding at runtime
Forwarding all necessary ports in the start script is not flexible
enough. There were also issues with the iptables-based networking in
Kubernetes (access to a service when logged into the host worked,
access from outside didn't).

As a solution for both, ssh gets started dynamically to forward ports
on demand. At the moment this is limited to TCP, but could also be
made more flexible as needed (UDP, Unix domain sockets).