QEMU: more flexible port forwarding
It's now possible to pass in anything as from/to specification that
"ssh -L" accepts. Also, the same ssh command can run some
command. This exposes some implementation detail, but turned out to be
very useful for starting the OIO CSI driver at the same time as
establishing port forwarding for it.