tracing: better logging (indention, responses/error, source location)
Previously, incoming and outgoing GRPC calls were logged like this:

  GRPC call: /csi.v0.Controller/CreateVolume
  GRPC request: name:"sanity" capacity_range:<required_bytes:2097152 limit_bytes:2097152 > volume_capabilities:<mount:<> access_mode:<mode:SINGLE_NODE_WRITER > >
  GRPC call: /oio.v0.Controller/ProvisionMallocBDev
  GRPC request: bdev_name:"sanity" size:2097152
  GRPC error: rpc error: code = AlreadyExists desc = Existing BDev sanity has wrong size 1048576
  GRPC error: rpc error: code = AlreadyExists desc = Existing BDev sanity has wrong size 1048576

It's not obvious at all why the same error gets printed twice (hint:
one is the result of the ProvisionMallocBDev call, the other of the
CreateVolume).

Now nested gRPC calls (= incoming call triggering one or more outgoing
calls) use indention to make the nesting more obvious. In addition,
only a single line is printed at the start of a call (shorter and
makes it obvious that the request data belongs to the call) and
printing the result repeats the method name:

  GRPC call: /csi.v0.Controller/CreateVolume name:"sanity" capacity_range:<required_bytes:2097152 limit_bytes:2097152 > volume_capabilities:<mount:<> access_mode:<mode:SINGLE_NODE_WRITER > >
    GRPC call: /oio.v0.Controller/ProvisionMallocBDev bdev_name:"sanity" size:2097152
    GRPC error: /oio.v0.Controller/ProvisionMallocBDev: rpc error: code = AlreadyExists desc = Existing BDev sanity has wrong size 1048576
  GRPC error: /csi.v0.Controller/CreateVolume: rpc error: code = AlreadyExists desc = Existing BDev sanity has wrong size 1048576

Logging calls that use the context end up with the same
indention. However, the varying length of the log prefix (includes
source code file name) makes that a bit less useful. While at it,
logging inside those calls now skips the helper function while going
up the stack, so the actual call location is shown.