e2e: better output logging by default
We can't expect every caller to remember to bump the glog verbosity
and redirect to stdout. It might not even be possible, for example
when invoking "go test" for multiple packages where some don't use
glog and thus don't know the flag. Therefore make the e2e testing
verbose by default.

With this change, we get to see gRPC logs in the OIO control
plane. Unfortunately we also get a lot of output from the E2E
framework :-/