e2e: run csi-test sanity tests against OIO CSI driver
The driver was already getting tested in local mode. When it runs
inside the virtual machine, we need the OIO control plane and the
sanity tests must connect to the driver through a forwarded socket.

The OIO control plane setup is the same as for the Kubernetes e2e
test, i.e. the OIO registry and controller service run inside the test
binary.