vendor: update SPDK source code
We can use the upstream repo again because the stub Go package
was added to master.

Updating SPDK fixes an issue where the vhost app locks up while
serving RPC requests.

The update also comes with an API change: constructing a BDev returns
a single name, not a list with one entry.