tracing: temporarily disable Jaeger
Jaeger depends on an unmaintained
component (https://github.com/jaegertracing/jaeger-lib/issues/32),
which isn't worth the risk considering that tracing is a) not a
required feature and b) depends on pending changes in the Kubernetes
sidecar containers to be more complete.

The code is commented out instead of removed because that way it may
be easier to bring it back (simply reverting a removal commit might
not apply cleanly).