Makefile: do not time out
When using "go test" to run the E2E test suite, the long-running test
gets aborted when using the default timeout. We don't know how long it
will run in total, so disable the time out entirely.

Running tests in a CI should use "ginkgo -p" because that can apply a
time out to each test case inside the suite and also runs a bit faster
because it parallelizes the suite.