Makefile, README.md: proxy support
When the usual env variables are set, we have to reconfigure Docker
inside the virtual machine and ensure that we pass the env variables
to curl and swupd when setting up the virtual machine.

We also need to ensure that the commands invoked during "docker build"
inherit the proxy env variables, otherwise installing the additional
packages fails.

0.0.0.0 needs special treatment and has to be in no_proxy, because Go
uses it as address for a local service.