Makefile: revise virtual machine image creation
By defining the concatenated shell command in a make variable that
gets appended to it becomes possible to interleave individual commands
and comments. The downside is that an embedded # gets interpreted by
make as starting a comment, so this has to be avoided. It was only
needed for creating the shell script, which is more readable anyway
when stored as a separate file that merely gets modified.

In terms of functionality there are also several improvements:
- QEMU gets properly killed now - it wasn't earlier because the
  co-process that got killed was the shell process running QEMU,
  not QEMU
- bash detects when the co-process dies and then unsets the
  env variables, which led to "ambiguous redirect" errors. Now
  we check whether QEMU still runs and print a better error message
  if not.
- shell tracing output gets suppressed for simple commands like "echo"
  to make the output more readable
- Clear Linux boot messages no longer get printed in the make output,
  which avoids the issue of changing console colors via the escape
  sequences used by Clear Linux