use tap networking
QEMU's userspace networking turned out to have a significant
performance impact. It is not entirely clear why, but as we will need
the ability to bring up multiple virtual machines on the same virtual
network, we might as well switch to such a setup now.

The downside is one additional tool dependency on the host (tunctl)
and one extra command that must be run manually. The hard-coded
network settings (specifically, 8.8.8.8 as DNS server and 192.168.7.0
as subnet) might not work for everyone.

virt-manager was considered, but that would pose its own challenges
and complexity and therefore was not tried.