doc: describe how to map from host name to controller
It is possible to use the compute node host name as unique ID for each
controller, as long as it is unique among all controllers connected to
the same registry. This is basically a simplified deployment scenario
for OIM. It has the advantage that the components on the host (like
OIM CSI driver) can simply use the local host name without additional
configuration or auto-detection of the controller ID.

The manual example uses this simplified approach.

More complex deployment scenarios might still be useful, so the
abstract concept of a "controller ID" is kept.