Makefile: VM with CRI-O, loopback networking, and no Weave
The goal was to avoid the startup cost for the Weave network
plugin. But in this configuration, pods like coredns and csi-pod fail
to start:

  Normal   Scheduled               45s   default-scheduler           Successfully assigned kube-system/coredns-78fcdf6894-jdx7f to kubernetes-master
  Warning  FailedCreatePodSandBox  44s   kubelet, kubernetes-master  Failed create pod sandbox: rpc error: code = Unknown desc = failed to get network status for pod sandbox k8s_coredns-78fcdf6894-jdx7f_kube-system_fc753603-bc14-11e8-907f-525400123456_0(d40f692fd4b7bcb00c84e9400326eab02e5f463520c70e31518c0b14c42b1f22): Unexpected command output Device "eth0" does not exist.
 with error: exit status 1

Disabling the network interface renaming didn't help.

This attempt merely gets committed to document the attempt.
The real solution probably will be to go back to Docker.