vendor-bom.csv: keep track of reviewed vendor components
We must ensure that only components with a compatible license get
imported. This is a manual process. Developers who add or remove
components must update the vendor-bom.csv file in the same commit to
document that they have done it. This also gives us an up-to-date
record of the incoming licenses.

The name of the components is based on how they are tracked at Intel.