multi-node demo cluster
This extends the virtual machine image creation so that it prepares a
cluster with four nodes. OIM can only be used on the first node,
because this is where the virtio-scsi device gets connected.

Therefore the Kubernetes node gets tagged with "intel.com/oim" and the
driver only gets deployed to that node via a daemon set. Right now,
applications must ensure that they get scheduled to that node by
specifying a node selector. This needs to be replaced with a mechanism
based on the CSI topology support; using that it should be possible to
schedule pods automatically onto nodes where OIM is supported.