test: deploy via manifest files
Instead of replicating the deployment with code, we can also load the
"deploy" manifest files. The utility code for doing this needs to go
upstream into test/e2e/manifest. Once that's done, the remaining code
in OIM will be a lot simpler.

While we are at it, registering the controller is now done by the
"virtual administrator" when setting up the cluster, instead of
relying on self-registration.