transport security and permission checks
All OIM connections must be secured with mutual TLS. The common name
is used to identify the peer. The registry accepts arbitrary peers,
but then checks the peer name to implement permission checks.

certstrap is used to generate certificates for testing. The same tests
as before pass, and negative tests for connection attempts that must
fail have been added to registry_test.go.