OIM CSI driver: flatten hierarchies, remove code
OIM CSI driver was derived from the hostpath driver, with a "generic"
default implementation and driver-specific overrides. In practice,
this "generic" implementation is now unmaintained upstream and then
not very helpful when forked in a driver.

Removing this additional complexity and directly implementing the
three CSI interfaces in the oimDriver struct makes it possible to
remove a lot of code and indirections.