test: remove broken QEMU unit tests
These tests were unintentionally (?) not included in "make test" (to
be fixed separately) and as a result of various other code changes no
longer build. A separate unit test for the test framework isn't really
needed (covered by running the tests), so instead of fixing the code
it gets removed.