update to Kubernetes 1.13/CSI 0.3/upstream ceph-csi, fix tests
Kubernetes 1.13 has been released, so we can use that instead of some
pre-1.13 master branch.

csi-test 0.3.5 no longer depends on the post-0.3 CSI spec, so plain
0.3 is fine now.

https://github.com/ceph/ceph-csi/pull/92 has been merged. We can use
the upstream release again. Because the 0.3 image tag is following the
master branch (https://github.com/ceph/ceph-csi/issues/96), we get the
latest features, which includes support for storing persistent data in
a config map (https://github.com/ceph/ceph-csi/pull/113).

That mode worked whereas storing on the node failed with an error
about not being able to create the file (probably because the
directory hadn't been created). Instead of trying to fix that, the
new feature is used.

Provisioning tests were failing because patching the driver name was
(no longer?) done correctly.