grpc: fix client common name check
When a server sets up TLS config with a peer name, the client's common
name was supposed to be checked at connection handshake time. Setting
ServerName` is not enough, a server has to validate the client by
providing a VerifyPeerCertificate function.

Of the OIM components, only OIM controller uses this because it only
expects connections from the OIM registry. Exactly that particular
testcase was missing.

OIM registry has to accept connections from different clients and then
does the check at the application level (i.e. in the different gRPC
server methods).