CSI 1.0 support
This moves as much code as possible into functions that are
independent of the CSI API, then implements both the old CSI 0.3 and
the new CSI 1.0 interfaces on top of that code.

A CSI driver can provide both CSI 0.3 and CSI 1.0 at the same time
under the traditional /var/lib/kubelet/plugins directory. But then it
is a bit uncertain which interface Kubernetes 1.14 will use. For OIM,
the approach with explicitly choosing between either 0.3 or 1.0 is
chosen to ensure that both interfaces are tested.

ceph-csi only supports CSI 0.3, so emulation of that is limited to
running also OIM CSI driver with CSI 0.3. oim-malloc uses CSI 1.0.

The newer CSI sidecar containers have finally released RBAC rules as
part of their source code. As part of moving to 1.0, the RBAC rules
get reduced to the bare minimum and copying from upstream is done via
scripts. "make update" will now ensure that the latest sidecar
containers and their RBAC rules are in the "deploy" directory.