test: use fixed sort order
The ordering of text files is locale-specific. When comparing against
the reference files we have to ensure that the same ordering is used
regardless of the environment and/or the actual "sort" implementation.

LC_ALL=C ensures that for GNU sort and glibc. LANG=C is added just as
an extra precaution for other implementations.