test: CNI plugins via symlinks
When switching to CRI-O, the flannel add-on failed because
/opt/cni/bin/loopback was not found.

The instructions for setting up CNI on the Clear Linux docs page are
inconsistent. What we do instead is symlinking into the right
directory. This is intentionally written such that it works at the
moment and also keeps working when Clear Linux itself creates
that directory.