clean up staging and target directory handling
The CSI spec requires the CO (= csi-sanity) to create the staging
directory and the *parent* of the target
directory. Kubernetes (https://github.com/kubernetes/kubernetes/issues/75535)
and csi-sanity have handled the target directory incorrectly by
creating it.

For sharing the staging directory to work, it must be exposed as a
volumeMount (https://github.com/kubernetes-csi/docs/issues/130). Then
OIM can rely on it being present and fail when it is missing.

The Kubernetes issue about the target directory is still open, so OIM
allows the CO to create it (but warns now) while creating it if
required.